Fix some issues when merging diff for fullscreen toggle
Some changes in indentation style caused some things to get lost when
getting the diff merged in since it couldn't be applied automatically.
This resulted in some extra widgets (I think) being made and showing up
incorrectly in the fullscreen mode.

This also adds a prompt on how to toggle the fullscreen mode and fixes a
typo in the ESC to quit prompt.