Add a default file output resolution and print a warning
The frameResolution defaults to (-1, -1) so if no resolution is
specified when trying to render to file we crash. This adds a default
resolution (1280, 720) to the render to file mode which is selected if
no resolution has been specified. Additionally a warning message is
printed out if no resolution was specified and we're choosing the
default one.