simplify SimpleAO, # ao samples and secondary ray length now set via params
The number of ao samples (previously set via a template parameter), now is
set via the "aoSamples" integer parameter. It defaults to 16 if not set.

The length of ao secondary rays (to brighten/darken ao shading) is set via
the "aoRayLength" float paramter. It defaults to 1e20f if not set.