Workaround crash on Windows
ISPC passes SIMD masks via pointer on Windows.
The *_bounds functions are implemented in ISPC but called by C++ code by
Embree, which did not pass a mask -- the mask is not used anyway,
because the functions operate with uniforms only.
However, the mask is still read, resulting in a crash when the pointer
is not properly aligned.
The workaround replaces the "if" with a conditional move, resulting in
the mask (and the crash) optimized away.