Correctly compute the global range of voxel values and give to worker 0
Worker 0 is the one who responds to ospGet* queries from the master so
we have each worker maintain its local range and in ospSetRegion do an
MPI_Reduce to find the current global min/max on the first worker. It
would be nice we we knew when the user was done calling ospSetRegion so
we didn't do a reduce for every call, but I'm not sure if there's a way
to do this.