Load the RAW volume file in 1.5GB chunks of datao
This fixes an incorrect assumption that taking a `dim.x x dim.y x 4`
chunk of the volume would be "small", e.g. on DNS which is
10240x7680x1536 these chunks are quite large. Instead we find the
largest convex piece of data we can fit in 1.5GB and read & set chunks
of data of that (handling remainder bits as well).

Cleaning up some comments that are now addressed

Clean up some of setting the global voxel range and use 1.5GB chunks

Next: support VOXELRANGE_IN_APP for the chunked raw loader (and RM?)
since that's the path forward with ospGet* being deprecated.