Don't open a window at all when `--nowin` is passed
Also added a `--max-accum` parameter to set the max number of
accumulations done for a frame and `--accum-reset` to instruct the
viewer to clear the framebuffer when accumID == maxAccum to continuously
render up to maxAccum accumulations then restart. This is useful for
benchmarking, e.g. where we'd want accumulation off and would run with
`--max-accum 1 --accum-reset`