Structured volumes can now use the scaleFactor param and upscale
The dimensions are assumed to be set to the scaled by dimensions by the
creator of the volume. You may need to get the param before commit is
called since we need to know the scale in setRegion. This is a bit
awkward since it breaks the commit finalizes stuff idea?