Working on the `ospFinalize` API to wrap up MPI in distrib rendering
**TODO:** The send/recv threads in the async messaging don't know that
they should quit so we crash out if they happen to make some more MPI
calls after we've called finalize and don't exit cleanly. So we need
a way of telling them to quit. Maybe they can check MPI_Finalized? Will
that add a lot of overhead to the system?