Adding a standard transfer function format for sample apps
The transfer functions also have names so integrating with
ospVolumeViewer is nicer and for safety I've put in a magic number to
identify functions. In case we want to change things without breaking
old files I've also added a version number to the header. Additionally
they can be used in anything that uses the volume scene parser from the
commandline parsers