Fix issue #63: Infinite loop when rendering isosurface at 0 isovalue
The ISPC rcp is maybe too approximate and returns -nan instead of
infinity in the case of 1/0. When we had an isosurface at isovalue 0 we
would end up potentially with sample == sample0 == 0.0 which would
produce rcpf(0.0) giving nan resulting in our loop in
RaycastVolumeRenderer_intersect testing against nan when checking if
it's exited the volume. As a result of the nan t value we wouldn't
accumulate any other colors or hit anything and would loop infinitely.