Add flag to log all frame times out as well
Passing -lft or --log-frame-times will log out the time taken to render
each frame, e.g. if an algorithm's performance changes over time this
can be used to make a plot of the data over the benchmark period instead
of just showing a single average fps for the whole run.