Fix inconsistencies in number of _ preceeding external name for creators
Everything but transfer functions and volumes used a double underscore,
e.g. `ospray_create_camera__##external_name()` while the volumes and
transfer functions used one underscore:
`ospray_create_volume_##external_name()`. This resulted in the migration
to the `OSP_REGISTER_OBJECT` macro breaking creation of volumes and
transferfunctions since they now would be with two underscores instead
of the one that we were looking for. I've switched to look for two
underscores for consistency with the other types.