Bad leaking hack to fix crash, need to put in the management class
The management class will be held by a shared_ptr which users will clone
to keep it alive for as long as they need it. So the MPIDevice, and
anyone else who relies on the mpi::send/recv working should hold
a copy of the shared ptr to the management class. So maybe instead
of going through free-methods and using globals/statics we should go
through this management class.