Fix not locking mpi on startup, template specializations not being found
This fixes the occasional hang when starting the mpi device due to not
properly locking the mpi initialization calls. This also fixes a bug in
the mpi device with the Intel compiler where the template
specializations for NewRenderer::runOnMaster and NewVolume::runOnMaster
were not being found, resulting in a crash.