Fix bug where master would fail to send accumulated tile to itself
After marking the tile->ownerID = -1 so we wouldn't re-send it
to ourself again in tileIsCompleted we would fail to set the ownerID
back to our id. As a result we would never send this tile to ourself
again, and would instead just mark it as being rendered (b/c it was
still rendered). So while the tile is rendered each frame, and
accumulated and so on, we just would never send the master write tile
messages for the following frames.