Check if our box is visible or not and send tile only if it is
Also we use this to check how many children will be coming too for this
tile, if we own it since we've got everyone else's boxes as well. TODO
next is to send this information on the background color tile so we
don't need the awkward logic of subtracting 1 from the box count only if
it was my child in there, or my N children subtract N. It will also fix
the background color being wrong.