Pass depth information for a tile on to the master
We use the MasterTileMessageBuilder to construct the message generically
on the send side, and then can just deserialize the POD structure on the
receiver. A bonus of this vs. the work items is we need less data copies
since we don't have the polymorphic structs we can't ship over the
network and can write directly to the message's buffer.