Use the void ptr command to receive a user-set world comm
Handle not duping the world com if we're partitioned,
dup is a collective, so if the ospray partition is a subset of the
world we'll hang on this collective. Instead, only dup the world com if
we're planning to use it as our communicator. Otherwise let the caller
of mpicommon::init configure the communicator as desired.