Add safe gather methods which take a byte stride
This is useful to support different layouts of the sphere geometries,
where the stride between colors may not be a multiple of floats (and
expressible by setting the color stride which was then interpreted as a
stride of floats). This essentially reverts back to the old semantics of
`color_stride` and `color_offset` where they refer to byte strides and
offsets. This also adds the `color_components` parameter for Spheres,
allowing users to explicitly set how many color components they have,
in case the OSPData holding the colors is not a `OSP_FLOAT3` or
`OSP_FLOAT4` array.