Testing a gather approach for the collection of finished master tiles
Avoid race condition where we would hang a tbb thread in the gather.
Occasionally the node would finish its work and see the frame as done,
however the TBB thread which noticed this was still tasked with
rendering some tiles needed by another node. The thread would then
enter the gather call to complete the frame and hang the processes, as
the other node would never receive the tile(s) it needed to finish the
frame.

Fix potential premature exits and race conditions when stopping

Essentially, do not stop the messaging layer. We don't know if the
messages we're receiving and processing could in turn spawn new
messages, and MAML may be quit prematurely in the case