Fixed memory leak in ospRemoveParam for ospNewData objects.
The specialization of ParametrizedObject::Param::set for
ManagedObjects takes care of calling refInc on a new object
and refDec on a previously existing object. However, the Any member
inside Param doesn't know that it has to call refDec on ManagedObject*,
and so the object is leaked when ospRemoveParam is called.

This fix gets the object and sets it to nullptr before removing.
A better fix would be to use a smart pointer for ManagedObjects in
ParametrizedObject::Param::set.