Correct epsilon computation bug when using a list of models
The Renderer parent class would reinterpret cast the OSPData passed as
"model" to a Model, then try to get its bounds and use this to adjust
the epsilon value, producing either 0s, nans, or 1e-6. This corrects
the issue by having the DistributedRaycastRenderer re-compute and
override this value found the by the Renderer parent class.