Update to frameop/tileop style more like geometries
The objects made through the API now act more like parameter staging
objects/factories which create the actual frame/tile lambda when they're
attached to a framebuffer. This makes it easier for ops to track
per-framebuffer state (e.g., as needed in the denoiser)

Added some semi-unused parent classes for tile/frame ops which we need
in the DFB for now to distinguish the frame ops so that we don't create
and attach them on the master. In the end I would probably remove these
once we can run the frame ops distributed?