fixing the non-physical behavior of the spot and point light sources
The color * intensity parametrization is now defined to represent the radian
intensity of the spot or point light source and is converted into radiance, if
radius > 0.
During the conversion the overall emitted power of the light source into the
scene stays the same when the size of the disk or sphere light changes.
Added legacy fix for spot lights (factor 0.5 on radiance) to keep the look
similar.
Added cosine falloff to real spot light (radius=0) to match disk light emission
behavior