Fix error in how DFB handled commits, do not clear the FB image
Previously the image would be cleared when the FB was committed, which
is not correct. It would also only be partially cleared, with the tile
operations recreated but the accumulation IDs not reset, leading to a
buggy output image until the FB was actually cleared. DFB::commit now
behaves like the LocalFB where the image/frame operations are reset but
the framebuffer is not cleared.