Refactor task management to launch tasks in ISPC
RenderTaskDesc now describes a region of the image to be rendered by a
thread and potentially stopped based on the adaptive rendering
termination criteria

- Inner parallel for launch moved to ISPC, runs over these render tasks
- Tile data stored in framebuffer instead of as stack vars to allocate
them ahead of time
- Fused render + write kernels for Renderers and PathTracer
- Sparse FB stores individual tiles for use in DFB and multidevice
- Refactor FB's to use fcn ptrs in ISPC for writing image data
- DFB data parallel: Store region's partial images in SparseFB layers
- Store FB channels in ISPC side memory

- SparseFB can have accumIDs without accum buffering
DFB needs the SparseFB to track accumIDs without accumulating into its
local data, since the accumulation must be done at the distributed level
during the tile operation. So we need to make a SparseFB without
accumulation buffering but with accumulation IDs

- Refactor tile ops to pixel ops to run during dispatch or as kernel
Decoupling the parallelism of per-pixel operations from the Tile size
gives us more flexibility in how/when to run these post-processing
operations in the pipeline.

- Multidevice Pixel ops are run by the subdevices in the sparse fb,
frame operations are not currently run in multidevice.
- Refactor separate ISPC/C++ Tile structs to shared Tile struct
- Update test image data for MPI distributed and readme about how to
  generate baseline images
- Sort tasks in z-order, SparseFB tasks are sorted locally in each tile

- Use thread local storage to reduce contention on progress atomic
This emulates the same reporting frequency we had before in OSPRay where
the outer parallel for over tiles would report progress, while the inner
loop over pixels did not report each individual pixel.