Add assigned data
ASSIGNED DATA

It is common for applications to need to massage data in a temporary
buffer before handing off to OSPRay. This change provides a way for
applications to transfer ownership of the buffer to try to avoid one
copy/delete cycle.

The additional arguments point to a custom deleter function (with
optional state if necessary) that OSPRay needs to call then it is done
with the buffer.

The C++ wrappers provide a cpp:MovedData(std::vector<>&) for the common
case.

TESTING

The new test_assigned_data test exercises the assigned data feature and
also demonstrate deleter functions as malloc and new counterparts. They
count and inspect the timeline of deletion calls to more fully validate
that the applications memory is in fact deleted by OSPRay through the
callback function.

Co-authored-by: Johannes Günther <johannes.guenther@intel.com>