Experimental support for device data
DEVICE DATA

For device data we now detect the case where the application shares that
with us and, by default, handle it by copying into an internal shared
buffer, just like we do for host data. The copy, as implemented,
requires a corresponding change in ispcrt that is expected in the next
release. In the interim the superbuild applies the needed commit as
patch.

The new OSPRAY_ALLOW_DEVICE_MEMORY environment flag turns on an
experimental mode where OSPRay instead leaves the buffer alone and
attempts to use the device data in place. The mode is experimental
because in most cases the app will simply crash as most code in OSPRay,
VKL and Embree do not yet support device data. Applications that know
exactly what they are doing ie using it only for VKL structured volumes
with the upcoming VKL release, or RGB surface colors, can use it to
avoid a copy.

TESTING

1) The --own-SYCL argument to ospTestSuite causes the test app to
   emulate in situ use cases where the sycl enabled application owns the
   context and shares it with ospray.

2) The new test_sharedusm_data test (only available when compiled with
   SYCL) creates a simple volumetric data set and validates that usm
   host, shared, or device data sharing work as expected.

Co-authored-by: Johannes Günther <johannes.guenther@intel.com>